CHANGED: Moved match arm code into separate module methods
